stream 9 down
was a little short, (tired, and its friday)

but we are ~33% done with stuff we want to do before we start heavily pushing this game to be played

with 9 real streams (10 if we count the short fall alseep one), and little over 3 hours each. we are at about 27 hours of work, halfway through a work week.

(I do work full time ontop of this, so I'm legit already exhausted by the time I start these)